requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic>=2.5.0
python-dotenv==1.0.0 
python-jose[cryptography]
python-dotenv